1. Fixed type for controller. 2. Removed swagger documentation. 3. Implemented find service by id method.

+++ src/main/java/com/rmm/services/serverapp/repository/ServiceRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * Handles service operations in a persistence storage.
  */
 @Repository
-public interface ServiceRepository extends CrudRepository<Service, Long> {
+public interface ServiceRepository extends CrudRepository<Service, Integer> {
 }
